requirements.txt
# MangaTranslator - Requirements File
# ====================================
# Dependencies for the multi-language comic translation system

# Core Libraries
numpy==1.24.2
opencv-python==4.9.0.80
pillow==10.3.0
requests>=2.25.0
python-dotenv>=0.19.0

# Machine Learning & AI
torch==2.5.0
ultralytics==8.1.43
transformers>=4.21.0
huggingface-hub==0.22.2
sentencepiece==0.2.0

# Translation Services
deep-translator==1.11.4
translators==5.9.1

# Multi-language OCR Engines
manga-ocr==0.1.11        # Japanese manga OCR (specialized)
paddleocr>=2.7.0         # Chinese manhua OCR (optimized)
easyocr>=1.7.0           # Korean manhwa + multilingual OCR

# OCR Support Libraries
shapely>=1.8.0           # Geometric operations for text detection
pyclipper>=1.3.0         # Polygon clipping for text regions

# Web Interface
gradio>=4.0.0            # Interactive web interface for the app
